Consolidate all tests under one project.
