+++ sdk/ai/Azure.AI.Projects/assets.json
@@ -2,5 +2,5 @@
   "AssetsRepo": "Azure/azure-sdk-assets",
   "AssetsRepoPrefixPath": "net",
   "TagPrefix": "net/ai/Azure.AI.Projects",
-  "Tag": "net/ai/Azure.AI.Projects_ac2e46e8e6"
+  "Tag": "net/ai/Azure.AI.Projects_0727c237a5"
 }
