+++ sdk/ai/Azure.AI.Projects/tests/AIAgentsTest.cs
@@ -7,6 +7,7 @@ using System;
 using System.Threading.Tasks;
 using Azure.AI.Agents.Persistent;
 using Microsoft.ClientModel.TestFramework;
+using Azure.AI.Projects.Tests.Utils;
 using NUnit.Framework;
 
 namespace Azure.AI.Projects.Tests;
