+++ sdk/ai/Azure.AI.Projects/tests/TestData/Agents/cua_browser_search.png
