+++ sdk/ai/Azure.AI.Projects/tests/TestData/Agents/cua_search_results.png
