+++ sdk/ai/Azure.AI.Projects/tests/TestData/Agents/tripadvisor_openapi.json
