+++ sdk/ai/Azure.AI.Projects/tests/TestData/Agents/weather_openapi.json
