+++ sdk/ai/Azure.AI.Projects/tests/Utils/HeaderTestPolicy.cs
@@ -5,7 +5,7 @@ using System.ClientModel.Primitives;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Azure.AI.Projects.Tests;
+namespace Azure.AI.Projects.Tests.Utils;
 
 internal class HeaderTestPolicy(IReadOnlyDictionary<string, string> _headers) : PipelinePolicy
 {
