--- sdk/ai/Azure.AI.Projects/tests_agents/Azure.AI.Projects.Agents.Tests.csproj
@@ -1,35 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworks>$(RequiredTargetFrameworks)</TargetFrameworks>
-    <!--Ignore XML doc comments on test types and members-->
-    <NoWarn>$(NoWarn);CS1591</NoWarn>
-    <!-- Package downgrade (BCL) -->
-    <NoWarn>$(NoWarn);NU1605</NoWarn>
-    <!-- Temporary fix to include latest changes from Projects. -->
-    <NoWarn>$(NoWarn);NU1105</NoWarn>
-    <!-- OPENAI001 - Ignore experimental warnings-->
-    <NoWarn>$(NoWarn);OPENAI001;SCME0001;</NoWarn>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\src\Azure.AI.Projects.csproj" />
-    <ProjectReference Include="..\..\Azure.AI.Projects.OpenAI\src\Azure.AI.Projects.OpenAI.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Azure.Identity" />
-    <PackageReference Include="NUnit" />
-    <PackageReference Include="NUnit3TestAdapter" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" />
-    <PackageReference Include="OpenTelemetry.Exporter.Console" />
-    <PackageReference Include="System.Linq.Async" />
-		<PackageReference Include="Microsoft.ClientModel.TestFramework" />
-    <PackageReference Include="Azure.Monitor.OpenTelemetry.Exporter" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<AssemblyAttribute Include="System.Reflection.AssemblyMetadataAttribute">
-			<_Parameter1>SourcePath</_Parameter1>
-			<_Parameter2>$(MSBuildProjectDirectory)</_Parameter2>
-		</AssemblyAttribute>
-	</ItemGroup>
-</Project>
