--- sdk/ai/Azure.AI.Projects/tests_agents/Utils/TestPipelinePolicy.cs
@@ -1,33 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using System;
-using System.ClientModel.Primitives;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace Azure.AI.Projects.Tests;
-
-internal partial class TestPipelinePolicy : PipelinePolicy
-{
-    private Action<PipelineMessage> _processMessageAction;
-
-    public TestPipelinePolicy(Action<PipelineMessage> processMessageAction)
-    {
-        _processMessageAction = processMessageAction;
-    }
-
-    public override void Process(PipelineMessage message, IReadOnlyList<PipelinePolicy> pipeline, int currentIndex)
-    {
-        _processMessageAction(message); // for request
-        ProcessNext(message, pipeline, currentIndex);
-        _processMessageAction(message); // for response
-    }
-
-    public override async ValueTask ProcessAsync(PipelineMessage message, IReadOnlyList<PipelinePolicy> pipeline, int currentIndex)
-    {
-        _processMessageAction(message); // for request
-        await ProcessNextAsync(message, pipeline, currentIndex);
-        _processMessageAction(message); // for response
-    }
-}
